% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMR.finder.R
\name{AMRfinder}
\alias{AMRfinder}
\title{AMRfinder: A function to perform segmentation analysis on DNA methylation data using the AMR method.}
\usage{
AMRfinder(
  intput_dat,
  y,
  cov.mod = NULL,
  controlist = list(maxdist = 300, maxseg = -1, mincpgs = 5, threads = 1, mode = 1, mtc =
    1, name = "sample", trend = 0.6, minNo = -1, minFactor = 0.8, valley = 0.7,
    minMethDist = 0.1, randomseed = 26061981)
)
}
\arguments{
\item{intput_dat}{A matrix containing input data with columns 'chr', 'pos', and CpG values.}

\item{y}{A vector or matrix representing the response variable.}

\item{cov.mod}{An optional data frame of covariates to include in the model.}

\item{controllist}{A list of control parameters}
}
\value{
A data frame with segmentation results including chr, start, end, #CpGs, correlation estimates, coefficients, p-values, methX, methY, and FDR.
}
\description{
AMRfinder: A function to perform segmentation analysis on DNA methylation data using the AMR method.
}
