% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMaxZ.R
\name{findMaxZC}
\alias{findMaxZC}
\title{Find Maximum Z Value}
\usage{
findMaxZC(s, t, mincpgs, XS)
}
\arguments{
\item{s}{Starting index (integer)}

\item{t}{Ending index (integer)}

\item{mincpgs}{Minimum number of CpGs (integer)}

\item{XS}{Two dimensional array/matrix}
}
\value{
A vector of two integers representing the indices with maximum Z value
}
\description{
Find Maximum Z Value
}
\examples{
\dontrun{
# Example usage
XS <- matrix(rnorm(100), ncol = 2)
result <- findMaxZC(1, 50, 3, XS)
}
}
